put the new flag into actual use and fix a bug
* put the new flag into actual use
* fix a bug that can not put items into the correct slot while importing container blocks
