+++ fastbuilder/bdump/blockNBT/pool.go
@@ -69,39 +69,3 @@ var SupportBlocksPool map[string]string = map[string]string{
 	"warped_hanging_sign":    "Sign",
 	// 告示牌
 }
-
-// 用于放置方块实体的通用接口
-type GeneralBlockNBT interface {
-	Decode() error     // 解析 NBT(map[string]interface{}) 数据为 GoLang Struct
-	WriteDatas() error // 放置方块并以正常玩家的行为写入方块实体数据
-}
-
-// 检查这个方块实体是否已被支持。
-// 如果尚未被支持，则返回空字符串，否则返回这种方块的类型。
-// 以告示牌为例，所有的告示牌都可以写作为 Sign
-func CheckIfIsEffectiveNBTBlock(blockName string) string {
-	value, ok := SupportBlocksPool[blockName]
-	if ok {
-		return value
-	}
-	return ""
-}
-
-// 取得用于放置目标方块实体的 接口/方法
-func GetMethod(pack Package) GeneralBlockNBT {
-	switch pack.Datas.Type {
-	case "CommandBlock":
-		return &CommandBlock{Package: &pack, NeedToPlaceBlock: true}
-		// 命令方块
-	case "Container":
-		return &Container{Package: &pack}
-		// 容器
-	case "Sign":
-		return &Sign{Package: &pack}
-		// 告示牌
-	default:
-		return &Default{Package: &pack}
-		// 其他尚且未被支持的方块实体
-	}
-	// 返回值
-}
