+++ fastbuilder/bdump/blockNBT/run.go
@@ -3,31 +3,9 @@ package blockNBT
 import (
 	"fmt"
 	GlobalAPI "phoenixbuilder/GameControl/GlobalAPI"
-	"phoenixbuilder/fastbuilder/mcstructure"
 	"phoenixbuilder/fastbuilder/types"
-	"strings"
 	"sync"
 )
-
-// 将 types.Module 转换为 blockNBT_depends.GeneralBlock
-func parseBlockModule(singleBlock *types.Module) (GeneralBlock, error) {
-	// init var
-	got, err := mcstructure.ParseStringNBT(singleBlock.Block.BlockStates, true)
-	if err != nil {
-		return GeneralBlock{}, fmt.Errorf("parseBlockModule: Could not parse block states; singleBlock.Block.BlockStates = %#v", singleBlock.Block.BlockStates)
-	}
-	blockStates, normal := got.(map[string]interface{})
-	if !normal {
-		return GeneralBlock{}, fmt.Errorf("parseBlockModule: The target block states is not map[string]interface{}; got = %#v", got)
-	}
-	// get block states
-	return GeneralBlock{
-		Name:   strings.Replace(strings.ToLower(strings.ReplaceAll(*singleBlock.Block.Name, " ", "")), "minecraft:", "", 1),
-		States: blockStates,
-		NBT:    singleBlock.NBTMap,
-	}, nil
-	// return
-}
 
 var apiIsUsing sync.Mutex
 
@@ -54,18 +32,30 @@ func PlaceBlockWithNBTData(
 	}
 	newRequest.Datas.StatesString = blockInfo.Block.BlockStates
 	newRequest.Datas.Position = [3]int32{int32(blockInfo.Point.X), int32(blockInfo.Point.Y), int32(blockInfo.Point.Z)}
-	newRequest.Datas.Type = CheckIfIsEffectiveNBTBlock(newRequest.Block.Name)
+	newRequest.Datas.Type = checkIfIsEffectiveNBTBlock(newRequest.Block.Name)
 	// get new request of place nbt block
-	placeBlockMethod := GetMethod(newRequest)
-	err = placeBlockMethod.Decode()
-	if err != nil {
-		return fmt.Errorf("PlaceBlockWithNBTData: Failed to place the entity block named %v at (%d,%d,%d), and the error log is %v", newRequest.Block.Name, blockInfo.Point.X, blockInfo.Point.Y, blockInfo.Point.Z, err)
+	var placeBlockMethod GeneralBlockNBT
+	if datas.Settings.AssignNBTData {
+		placeBlockMethod = getMethod(newRequest)
+		err = placeBlockMethod.Decode()
+		if err != nil {
+			return fmt.Errorf("PlaceBlockWithNBTData: Failed to place the entity block named %v at (%d,%d,%d), and the error log is %v", newRequest.Block.Name, blockInfo.Point.X, blockInfo.Point.Y, blockInfo.Point.Z, err)
+		}
+		// if the user wants us to assign NBT data
+	} else {
+		if newRequest.Datas.Type == "CommandBlock" {
+			placeBlockMethod = &CommandBlock{Package: &newRequest, NeedToPlaceBlock: true}
+		} else {
+			placeBlockMethod = &Default{Package: &newRequest}
+		}
+		// uf the user does not want us to assign NBT data
 	}
+	// get method and decode nbt data into golang struct
 	err = placeBlockMethod.WriteDatas()
 	if err != nil {
 		return fmt.Errorf("PlaceBlockWithNBTData: Failed to place the entity block named %v at (%d,%d,%d), and the error log is %v", newRequest.Block.Name, blockInfo.Point.X, blockInfo.Point.Y, blockInfo.Point.Z, err)
 	}
-	// place block with nbt datas
+	// assign nbt data
 	return nil
 	// return
 }
